Bouncy256k1: move static conversion methods from BC, remove BC
